Move to a hook.
