+++ src/components/PackageRow.tsx
@@ -1,42 +1,17 @@
 import React from 'react';
 import { Box } from 'ink';
 import { Style } from '@boost/cli';
-import Artifact from '../Artifact';
-import BundleArtifact from '../BundleArtifact';
 import Package from '../Package';
 import ArtifactList from './ArtifactList';
 import Environment from './Environment';
+import useGroupedArtifacts from './hooks/useGroupedArtifacts';
 
 export interface PackageRowProps {
   package: Package;
 }
 
 export default function PackageRow({ package: pkg }: PackageRowProps) {
-  const ungrouped: Artifact[] = [];
-  let groups: Record<string, Set<Artifact>> = {};
-
-  // Group artifacts by platform and support
-  pkg.artifacts.forEach((artifact) => {
-    if (artifact instanceof BundleArtifact) {
-      artifact.builds.forEach((build) => {
-        const key = `${build.platform}:${build.support}`;
-        const set = groups[key] || new Set();
-
-        set.add(artifact);
-        groups[key] = set;
-      });
-    } else {
-      ungrouped.push(artifact);
-    }
-  });
-
-  // If only 1 group, collapse into a single target
-  const envs = Object.keys(groups);
-
-  if (envs.length === 1) {
-    ungrouped.unshift(...Array.from(groups[envs[0]]));
-    groups = {};
-  }
+  const { envs, groups, ungrouped } = useGroupedArtifacts(pkg);
 
   return (
     <Box flexDirection="column" marginTop={1}>
